requirements.txt
# Django
Django==1.10.6

# Tastypie
python-mimeparse==1.6.0
python-dateutil==2.6.0
django-tastypie==0.13.3

